# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateAccountRequest:
    properties:
      account_description:
        type: string
      account_name:
        type: string
      currency:
        $ref: '#/definitions/models.Currency'
    type: object
...
  models.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  models.Currency:
    enum:
    - USD
    - EUR
    - GBP
    type: string
    x-enum-varnames:
    - USD
    - EUR
    - GBP
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Banking System API.
  title: Banking System API
  version: "1.0"
paths:
  /account/create:
    post:
      consumes:
      - application/json
      description: Create a new account for a user with specified currency
      parameters:
      - description: Create account request
        in: body
        name: createAccountRequest
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountRequest'
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: integer
                  type: object
              type: object
        "400":
          description: Invalid request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new account
      tags:
      - account
  /account/delete:
    delete:
      consumes:
      - application/json
      description: Permanently delete an account and all associated data
      parameters:
      - description: Account ID
        in: query
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account deleted successfully
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Account not found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an account
      tags:
      - account
  /account/get:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific account
      parameters:
      - description: Account ID
        in: query
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account details retrieved successfully
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Unauthorized access to account
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get an account by ID
      tags:
      - account
  /account/get-accounts:
    get:
      consumes:
      - application/json
      description: Get all accounts for a user with filtering, sorting, and pagination
      parameters:
      - description: Minimum balance filter
        in: query
        name: minBalance
        type: number
      - description: Maximum balance filter
        in: query
        name: maxBalance
        type: number
      - description: Currency filter (USD, EUR, GBP)
        in: query
        name: currency
        type: string
      - description: Date from filter (RFC3339)
        in: query
        name: dateFrom
        type: string
      - description: Date to filter (RFC3339)
        in: query
        name: dateTo
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Sort field (balance, currency, created_at)
        in: query
        name: sortField
        type: string
      - description: Sort direction (ASC, DESC)
        in: query
        name: sortDirection
        type: string
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all accounts for a user
      tags:
      - account
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user with email and password
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login a user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Logout a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"